KAFKA-10310: Interfaces for SnapshotReader and SnapshotWriter
