+++ core/src/test/scala/integration/kafka/server/RaftClusterSnapshotTest.scala
@@ -23,7 +23,7 @@ import kafka.testkit.TestKitNodes
 import kafka.utils.TestUtils
 import org.apache.kafka.common.utils.BufferSupplier;
 import org.apache.kafka.metadata.MetadataRecordSerde
-import org.apache.kafka.snapshot.SnapshotReader
+import org.apache.kafka.snapshot.SnapshotReaderImpl
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotEquals
 import org.junit.jupiter.api.Assertions.assertTrue
@@ -76,7 +76,7 @@ class RaftClusterSnapshotTest {
       // For every controller and broker perform some sanity checks against the lastest snapshot
       for ((_, raftManager) <- cluster.raftManagers().asScala) {
         TestUtils.resource(
-          SnapshotReader.of(
+          SnapshotReaderImpl.of(
             raftManager.replicatedLog.latestSnapshot.get(),
             new MetadataRecordSerde(),
             BufferSupplier.create(),
