+++ core/src/test/scala/unit/kafka/server/metadata/BrokerMetadataSnapshotterTest.scala
@@ -30,7 +30,7 @@ import org.apache.kafka.metadata.MetadataRecordSerde
 import org.apache.kafka.queue.EventQueue
 import org.apache.kafka.raft.OffsetAndEpoch
 import org.apache.kafka.server.common.ApiMessageAndVersion
-import org.apache.kafka.snapshot.{MockRawSnapshotWriter, SnapshotWriter}
+import org.apache.kafka.snapshot.{MockRawSnapshotWriter, SnapshotWriter, SnapshotWriterImpl}
 import org.junit.jupiter.api.Assertions.{assertEquals, assertFalse, assertTrue}
 import org.junit.jupiter.api.Test
 
@@ -50,7 +50,7 @@ class BrokerMetadataSnapshotterTest {
                        committedEpoch: Int,
                        lastContainedLogTime: Long): SnapshotWriter[ApiMessageAndVersion] = {
       val offsetAndEpoch = new OffsetAndEpoch(committedOffset, committedEpoch)
-      SnapshotWriter.createWithHeader(
+      SnapshotWriterImpl.createWithHeader(
         () => Optional.of(new MockRawSnapshotWriter(offsetAndEpoch, consumeSnapshotBuffer)),
         1024,
         MemoryPool.NONE,
