+++ metadata/src/test/java/org/apache/kafka/controller/QuorumControllerTest.java
@@ -75,6 +75,7 @@ import org.apache.kafka.raft.Batch;
 import org.apache.kafka.server.common.ApiMessageAndVersion;
 import org.apache.kafka.snapshot.RawSnapshotReader;
 import org.apache.kafka.snapshot.SnapshotReader;
+import org.apache.kafka.snapshot.SnapshotReaderImpl;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 
@@ -400,7 +401,7 @@ public class QuorumControllerTest {
     }
 
     private SnapshotReader<ApiMessageAndVersion> createSnapshotReader(RawSnapshotReader reader) {
-        return SnapshotReader.of(
+        return SnapshotReaderImpl.of(
             reader,
             new MetadataRecordSerde(),
             BufferSupplier.create(),
