+++ raft/src/main/java/org/apache/kafka/snapshot/SnapshotReader.java
@@ -18,15 +18,8 @@
 package org.apache.kafka.snapshot;
 
 import java.util.Iterator;
-import java.util.NoSuchElementException;
-import java.util.Optional;
-import java.util.OptionalLong;
-
-import org.apache.kafka.common.utils.BufferSupplier;
 import org.apache.kafka.raft.Batch;
 import org.apache.kafka.raft.OffsetAndEpoch;
-import org.apache.kafka.server.common.serialization.RecordSerde;
-import org.apache.kafka.raft.internals.RecordsIterator;
 
 /**
  * A type for reading an immutable snapshot.
@@ -43,119 +36,29 @@ import org.apache.kafka.raft.internals.RecordsIterator;
  * offsets and epoch from the log are included in this snapshot. Both of these values are
  * inclusive.
  */
-public final class SnapshotReader<T> implements AutoCloseable, Iterator<Batch<T>> {
-    private final OffsetAndEpoch snapshotId;
-    private final RecordsIterator<T> iterator;
-
-    private Optional<Batch<T>> nextBatch = Optional.empty();
-    private OptionalLong lastContainedLogTimestamp = OptionalLong.empty();
-
-    private SnapshotReader(
-        OffsetAndEpoch snapshotId,
-        RecordsIterator<T> iterator
-    ) {
-        this.snapshotId = snapshotId;
-        this.iterator = iterator;
-    }
-
+public interface  SnapshotReader<T> extends AutoCloseable, Iterator<Batch<T>> {
     /**
      * Returns the end offset and epoch for the snapshot.
      */
-    public OffsetAndEpoch snapshotId() {
-        return snapshotId;
-    }
+    OffsetAndEpoch snapshotId();
 
     /**
      * Returns the last log offset which is represented in the snapshot.
      */
-    public long lastContainedLogOffset() {
-        return snapshotId.offset - 1;
-    }
+    long lastContainedLogOffset();
 
     /**
      * Returns the epoch of the last log offset which is represented in the snapshot.
      */
-    public int lastContainedLogEpoch() {
-        return snapshotId.epoch;
-    }
+    int lastContainedLogEpoch();
 
     /**
      * Returns the timestamp of the last log offset which is represented in the snapshot.
      */
-    public long lastContainedLogTimestamp() {
-        if (!lastContainedLogTimestamp.isPresent()) {
-            nextBatch.ifPresent(batch -> {
-                throw new IllegalStateException(
-                    String.format(
-                        "nextBatch was present when last contained log timestamp was not present",
-                        batch
-                    )
-                );
-            });
-            nextBatch = nextBatch();
-        }
-
-        return lastContainedLogTimestamp.getAsLong();
-    }
-
-    @Override
-    public boolean hasNext() {
-        if (!nextBatch.isPresent()) {
-            nextBatch = nextBatch();
-        }
-
-        return nextBatch.isPresent();
-    }
-
-    @Override
-    public Batch<T> next() {
-        if (!hasNext()) {
-            throw new NoSuchElementException("Snapshot reader doesn't have any more elements");
-        }
-
-        Batch<T> batch = nextBatch.get();
-        nextBatch = Optional.empty();
-
-        return batch;
-    }
+    long lastContainedLogTimestamp();
 
     /**
      * Closes the snapshot reader.
      */
-    public void close() {
-        iterator.close();
-    }
-
-    public static <T> SnapshotReader<T> of(
-        RawSnapshotReader snapshot,
-        RecordSerde<T> serde,
-        BufferSupplier bufferSupplier,
-        int maxBatchSize
-    ) {
-        return new SnapshotReader<>(
-            snapshot.snapshotId(),
-            new RecordsIterator<>(snapshot.records(), serde, bufferSupplier, maxBatchSize)
-        );
-    }
-
-    /**
-     * Returns the next non-control Batch
-     */
-    private Optional<Batch<T>> nextBatch() {
-        while (iterator.hasNext()) {
-            Batch<T> batch = iterator.next();
-
-            if (!lastContainedLogTimestamp.isPresent()) {
-                // The Batch type doesn't support returning control batches. For now lets just use
-                // the append time of the first batch
-                lastContainedLogTimestamp = OptionalLong.of(batch.appendTimestamp());
-            }
-
-            if (!batch.records().isEmpty()) {
-                return Optional.of(batch);
-            }
-        }
-
-        return Optional.empty();
-    }
+    void close();
 }
